2013-07-09  Dario Rodriguez <dario@artica.es>
	* integria_db.sql: Fixed a mistake in tinventory_track table.
	* include/functions_incidents.php: Moved "incidents_check_incident_acl"
	function to enterprise and adapted other funciton to new ubication.

+++ integria_db.sql
@@ -988,7 +988,7 @@ CREATE TABLE `tholidays` (
 
 CREATE TABLE `tinventory_track` (
   `id` int(10) unsigned NOT NULL auto_increment,
-  `id_inventory` bigint(20) unsigned NOT NULL default '0',
+  `id_inventory` mediumint(8) unsigned NOT NULL default '0',
   `id_aditional` varchar(60) NOT NULL DEFAULT '0',
   `state` int(10) unsigned NOT NULL default '0',
   `timestamp` datetime NOT NULL default '0000-00-00 00:00:00',
